+++ LegendWeathers/Weathers/BloodMoonWeather.cs
@@ -1,24 +1,30 @@
-﻿namespace LegendWeathers.Weathers
+﻿using WeatherRegistry;
+
+namespace LegendWeathers.Weathers
 {
     public class BloodMoonWeather : LegendWeather
     {
         public static string weatherAlert = "";
+
+        public static ImprovedWeatherEffect? BloodMoonEffectReference { get; private set; } = null;
 
         public BloodMoonWeather() : base(Plugin.instance.bloodMoonDefinition) { }
 
         public override void OnEnable()
         {
             base.OnEnable();
-            if (!WeatherRegistry.WeatherManager.IsSetupFinished)
+            if (!WeatherManager.IsSetupFinished)
                 return;
+            BloodMoonEffectReference = ConfigHelper.ResolveStringToWeather("bloodmoon").Effect;
             EnableVanillaSun(false);
         }
 
         public override void OnDisable()
         {
             base.OnDisable();
-            if (!WeatherRegistry.WeatherManager.IsSetupFinished)
+            if (!WeatherManager.IsSetupFinished)
                 return;
+            BloodMoonEffectReference = null;
             EnableVanillaSun(true);
         }
     }